Handle the dialog that is presented on first try to communicate with a `spillo:///` URL
When running this extension the first time (or every time, when the user doesn't choose 'always allow' in the dialog) Firefox pops up a request if the user wants to allow the extension to communicate with the Spillo native binary.

In the current solution this will simply lead to the whole construct not working, because we're closing the tab on the user after 100ms, so they can't even see that request, much less acknowledge it.

While trying to work around that with various feature of the tabs API (all of which didn't work) I've finally realized, that we can get rid of the whole new Tab dance altogether for good and solve that problem in the same go.

So now we simply append an invisible link (invisible, because it contains no text), make the browser click it and invoke Spillo as the protocol handler that way.
This also has the benefit, that the permissions dialog (if still shown) is presented in the context of the current tab, so the user can't really miss it.